+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### v0.0.41
 
 - Add Note list item component
+- Add SelectPicker Required Prop
+
+--
 
 ### v0.0.40